private_cloud/requirements-private.txt
pycryptodome
langchain
pyyaml
